requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
requests
pydantic
passlib[bcrypt]
bcrypt>=4.0.0,<4.1.0
python-jose[cryptography]
